making the 386 code somewhat follow the stupid rules
lol, I can't believe that on the actual bootstrap code people
have to put up with this petty picky shitty rules that some
asshole come up with who has an ego way past the douchebag
threshold.

I am so so so glad I'm running my own fork of this with the
"does it work in the browser? good. let's roll" coding standard.